+++ internal/applets/applet.go
@@ -82,9 +82,11 @@ import (
 	"github.com/nao1215/mimixbox/internal/applets/netutils/nc"
 	"github.com/nao1215/mimixbox/internal/applets/netutils/ping"
 	"github.com/nao1215/mimixbox/internal/applets/netutils/whris"
+	"github.com/nao1215/mimixbox/internal/applets/securityutils/pwcrack"
 	"github.com/nao1215/mimixbox/internal/applets/securityutils/pwgen"
 	"github.com/nao1215/mimixbox/internal/applets/securityutils/pwscore"
 	"github.com/nao1215/mimixbox/internal/applets/securityutils/unshadow"
+	"github.com/nao1215/mimixbox/internal/applets/securityutils/zippwcrack"
 
 	//"github.com/nao1215/mimixbox/internal/applets/loginutils/chsh"
 	validShell "github.com/nao1215/mimixbox/internal/applets/debianutils/valid-shell"
@@ -284,6 +286,7 @@ func init() {
 		"pwgen":            reg(pwgen.New()),
 		"pwscore":          reg(pwscore.New()),
 		"printf":           reg(printf.New()),
+		"pwcrack":          reg(pwcrack.New()),
 		"pwd":              reg(pwd.New()),
 		"readlink":         reg(readlink.New()),
 		"realpath":         reg(realpath.New()),
@@ -342,6 +345,7 @@ func init() {
 		"xargs":            reg(xargs.New()),
 		"xxd":              reg(xxd.New()),
 		"zip":              reg(zipCmd.New()),
+		"zip-pwcrack":      reg(zippwcrack.New()),
 		"who":              reg(who.New()),
 		"whoami":           reg(whoami.New()),
 		"whris":            reg(whris.New()),