+++ test/it/spec/pwcrack_spec.sh
@@ -1,0 +1,10 @@
+Describe 'pwcrack'
+    Include securityutils/pwcrack_test.sh
+    BeforeEach 'Setup'
+    AfterEach 'CleanUp'
+    It 'finds a weak password in the wordlist'
+        When call TestPwcrack
+        The output should include ': secret'
+        The status should be success
+    End
+End